+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## 0.1.8
+
+- (ios) Fix bridging header import
+- (android) Don't launch custom call activity on Answer / Decline actions in displayCustomCall
+
+## 0.1.7
+
+- (ios) Fix podspec name
+
+## 0.1.6
+
+- Allow setting ringtone uri on custom notification (#6)
+
+## 0.1.5
+
+- Add ability to customize heads-up notification (#5)
+
+## 0.1.4
+
+- Add `dismissCustomIncomingCall` function (#4)
+
+## 0.1.3
+
+- Add `displayCustomIncomingCall` function (#3)
+
 ## 0.1.2
 
 - Add `isCurrentDeviceSupported` property